Refactored the Diffing Engine as needed
