+++ Diffing_Engine/Compute/Diffing.cs
@@ -118,55 +118,22 @@ namespace BH.Engine.Diffing
 
         }
 
+
+        /***************************************************/
+
+
         [Description("Dispatch objects in two sets into the ones exclusive to one set, the other, or both.")]
         [Input("setA", "A previous version of a Stream")]
         [Input("setB", "A new version of a Stream")]
         [Input("diffConfig", "Sets configs such as properties to be ignored in the diffing, or enable/disable property-by-property diffing.")]
-        [MultiOutputAttribute(0, "OnlySetA", "Object existing exclusively in setA")]
-        [MultiOutputAttribute(1, "Intersection-A", "Objects existing in both sets. Returns objects originally from set A.")]
-        [MultiOutputAttribute(2, "Intersection-B", "Objects existing in both sets. Returns objects originally from set B.")]
-        [MultiOutputAttribute(3, "OnlySetB", "Object existing exclusively in setB")]
-        public static Output<List<IBHoMObject>, List<IBHoMObject>, List<IBHoMObject>, List<IBHoMObject>> Diffing(IEnumerable<IBHoMObject> setA, IEnumerable<IBHoMObject> setB, DiffConfig diffConfig = null)
+        public static Delta Diffing(IEnumerable<IBHoMObject> setA, IEnumerable<IBHoMObject> setB, DiffConfig diffConfig = null, bool saveHashInTags = true)
         {
             Stream streamA = BH.Engine.Diffing.Create.Stream(setA, diffConfig, "");
             Stream streamB = BH.Engine.Diffing.Create.Stream(setB, diffConfig, "");
 
-            VennDiagram<IBHoMObject> diagram = Engine.Data.Create.VennDiagram(streamA.Objects, streamB.Objects, new HashFragmComparer<IBHoMObject>());
-
-            return new Output<List<IBHoMObject>, List<IBHoMObject>, List<IBHoMObject>, List<IBHoMObject>>
-            {
-                Item1 = diagram.OnlySet1,
-                Item2 = diagram.Intersection.Select(tuple => tuple.Item1).ToList(),
-                Item3 = diagram.Intersection.Select(tuple => tuple.Item2).ToList(),
-                Item4 = diagram.OnlySet2,
-            };
+            return Diffing(streamA, streamB);
         }
 
 
-        /***************************************************/
-
-
-        /***************************************************/
-        /**** Public Fields                             ****/
-        /***************************************************/
-
-        public class HashFragmComparer<T> : IEqualityComparer<T> where T : IBHoMObject
-        {
-            public bool Equals(T x, T y)
-            {
-                string firstHash = x.GetHashFragment().Hash;
-                string secHash = y.GetHashFragment().Hash;
-                if (!string.IsNullOrEmpty(firstHash) && !string.IsNullOrEmpty(secHash) && firstHash == secHash)
-                    return true;
-                else
-                    return false;
-            }
-            public int GetHashCode(T obj)
-            {
-                return obj.GetHashFragment().GetHashCode();
-            }
-        }
-
-        /***************************************************/
     }
 }
