+++ Diffing_Engine/Diffing_Engine.csproj
@@ -74,6 +74,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Compute\HashComparing.cs" />
     <Compile Include="Compute\Diffing.cs" />
     <Compile Include="Compute\DiffingHash.cs" />
     <Compile Include="Convert\ToDiffingByteArray.cs" />
